Update documentation
I moved the example to the top, so people taking a first look can immediatley see how it might be used.
Furthermore, functions that are mainly used internally are gone from the documentation for now. They just added a lot of noise for probably not having been used by anybody else.